move into schema.py, add docs
